+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/TypeNode.java
@@ -65,4 +65,8 @@ public abstract class TypeNode extends DelphiNode implements Typed {
 
   @NotNull
   protected abstract Type createType();
+
+  public void clearCachedType() {
+    this.type = null;
+  }
 }