+++ src/main/java/org/sonar/plugins/delphi/type/DelphiClassReferenceType.java
@@ -1,19 +1,21 @@
 package org.sonar.plugins.delphi.type;
 
+import static org.sonar.plugins.delphi.symbol.scope.UnknownScope.unknownScope;
+
 import org.jetbrains.annotations.NotNull;
 import org.sonar.plugins.delphi.symbol.scope.DelphiScope;
 import org.sonar.plugins.delphi.type.Type.ClassReferenceType;
 import org.sonar.plugins.delphi.type.Type.ScopedType;
 
 public class DelphiClassReferenceType extends DelphiType implements ClassReferenceType, ScopedType {
-  private final ScopedType classType;
+  private Type classType;
 
-  private DelphiClassReferenceType(ScopedType classType) {
+  private DelphiClassReferenceType(Type classType) {
     super("class of " + classType.getImage());
     this.classType = classType;
   }
 
-  public static ScopedType classOf(ScopedType classType) {
+  public static ScopedType classOf(Type classType) {
     return new DelphiClassReferenceType(classType);
   }
 
@@ -25,11 +27,16 @@ public class DelphiClassReferenceType extends DelphiType implements ClassReferen
   @Override
   @NotNull
   public DelphiScope typeScope() {
-    return classType.typeScope();
+    return classType instanceof ScopedType ? ((ScopedType) classType).typeScope() : unknownScope();
   }
 
   @Override
-  public ScopedType classType() {
+  public Type classType() {
     return classType;
+  }
+
+  @Override
+  public void setClassType(Type type) {
+    this.classType = type;
   }
 }