+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improve progress reporting.
+- Improve type resolution around implicit forward declarations.
 - Pointer names are now expected to take the dereferenced type name and replace a leading `T` with a
   `P` in `PointerNameRule`.